refactor(*): refactored assets apis
