+++ pkg/assets/assets.go
@@ -1,7 +1,10 @@
 package assets
 
-import "github.com/eucrypt/unmarshal-go-sdk/pkg/constants"
+import (
+	"github.com/eucrypt/unmarshal-go-sdk/pkg/assets/types"
+	"github.com/eucrypt/unmarshal-go-sdk/pkg/constants"
+)
 
 type Assets interface {
-	GetAssets(chain constants.Chain, address string) (response AssetDetailsV1Resp, err error)
+	GetAssets(chain constants.Chain, address string) (response types.AssetDetailsV1Resp, err error)
 }
