+++ pkg/assets/assets_store_impl.go
@@ -1,6 +1,7 @@
 package assets
 
 import (
+	"github.com/eucrypt/unmarshal-go-sdk/pkg/assets/types"
 	"github.com/eucrypt/unmarshal-go-sdk/pkg/constants"
 	"github.com/eucrypt/unmarshal-go-sdk/pkg/session"
 	"strings"
@@ -10,18 +11,16 @@ type V1Store struct {
 	sess session.Session
 }
 
-const V1Path = "v1"
-
 func New(sess session.Session) V1Store {
 	return V1Store{sess: sess}
 }
 
-func (a V1Store) GetAssets(chain constants.Chain, address string) (response AssetDetailsV1Resp, err error) {
-	if !constants.GetAssets.SupportsChain(chain) {
+func (a V1Store) GetAssets(chain constants.Chain, address string) (response types.AssetDetailsV1Resp, err error) {
+	if !constants.Assets_GetAssets.SupportsChain(chain) {
 		return response, constants.UnsupportedChainError
 	}
-
-	path := strings.Join([]string{V1Path, chain.String(), "address", address, "assets"}, "/")
+	path := strings.Replace(constants.Assets_GetAssets.GetURI(), ":chain", chain.String(), 1)
+	path = strings.Replace(path, ":address", address, 1)
 	err = a.sess.Client.Get(&response, path, nil)
 	return
 }
